requirements.txt
flask >= 2.0
Pillow >= 8.0
whatimage
pyheif >= 0.5
#zipfile
#tarfile